+++ src/main/java/com/github/shk0da/goldendragon/repository/PricesRepository.java
@@ -8,6 +8,6 @@ import java.util.Map;
  * book data by ticker.
  */
 public class PricesRepository
-        extends AbstractRepository<TickerInfo.Key, Map<String, Map<Double, Integer>>> {
+        extends AbstractRepository<TickerInfo.Key, Map<String, Map<Double, Long>>> {
     public static final PricesRepository INSTANCE = new PricesRepository();
 }
